Remove tenant concept, rename getTenant util
Renames getTenant.tsx to getAppConfig.tsx and removes the backward-compatible getTenant alias. Drops the tenant field from T_Config, T_HistoryEntry, config.json files, and API responses. Removes the unused T_Tenant type and the free/config.json config. Updates LinkedIn prompt copy to drop multi-tenant references.
