+++ apps/admin/public/nxadmin/config.json
@@ -1,5 +1,4 @@
 {
-    "tenant": "admin",
     "siteName": "NX°",
     "description": "Level Five AIs only",
     "url": "https://admin.goldlabel.pro/",
